Fixes for Issue #1
Line 43: By default the background color had an alpha of 0. Occluder would not work and needs an alpha of 1
Line 81: Typo in "_Color" was "_Color,"
Line 181: By default the RenderPass is set to "Before Rendering Pre Passes", to me it works only if it sets at minimum "After rendering Opaques"
Line 185 & 186: Without this, the render feature throw errors as the settings are not created beforehand.